requirements.txt
discord.py
discord
aiosqlite
python-dotenv
asyncio